changed int output into float output output into

+++ HomeAssigment-2/HomeAssigment-2B/headTb.h
@@ -13,4 +13,4 @@ int GetInt(const char* arr, int* n);
 
 char* ReadFile(std::string name);
 
-int Poland(const char* arr, int size);
+float Poland(const char* arr, int size);
